unmarshal: always overwrite duplicate keys in target maps
This change introduces a difference in parsing from the reference
parser: the reference parser will always choose the _first_ value seen
associated with a key, while we will always choose the _last_.

This is not regarded as a compatibility issue, as there is no way to
provide a useful and re-serializable value in the presence of duplicate
keys.

Fixes https://github.com/DHowett/go-plist/issues/37.