Implement sparse/keyed array support
The following is a valid plist XML:

    <array>
        <key>100</key>
        <string>foo</string>
        <key>200</key>
        <string>bar</string>
    </array>

This enables parsing such arrays to dictionaries, which can in turn be
unmarshaled to Go maps, similarly to <dict> elements.

This isn't great, as ideally keys should be integers, and order should
be preserved - but at least it allows for some kind of parse.